Minor: viewport pitch = 0 for certain layers
